move test models into conftest and re-organize helpers into conftest
